moved doing.go -> context.go
